Creating a new run will now use the same plan actualDistance logic as creating/updating a plan. Fixes bug where week.actualDistance was only updated on days where the week started on (Mondays)
